Moving renderCurve into one separately file
